moved token validation to new file. login /users validates and retrieves all users
